remove old files, update src/, add dynamic width
