+++ src/lib/components/mediaPreview/MediaAudioPreview.svelte
@@ -1,22 +1,12 @@
 <script lang="ts">
-  export let options: Omit<Media, "isSelected">;
+  import LoadIcon from "../util/LoadIcon.svelte";
 
-  let audioContext = new AudioContext();
+  export let metadata: StudioMediaMetadata;
 
-  fetch(options.src).then((res) => res.arrayBuffer().then((buffer) => audioContext.decodeAudioData(buffer, onDecode)));
+  let canvas: HTMLCanvasElement;
 
-  const onDecode = (buffer: AudioBuffer) => {
-    // create buffer from loaded buffer
-    const source = audioContext.createBufferSource();
-    source.buffer = buffer;
-    source.connect(audioContext.destination);
-
-    // copy buffer data to float32Array
-    const dataArray = new Float32Array(source.buffer.length);
-    source.buffer.copyFromChannel(dataArray, 0);
-
-    // # steps per pixel on screen
-    const step = Math.ceil(dataArray.length / canvas.width);
+  metadata.audioData.then((data) => {
+    const step = Math.ceil(data.length / canvas.width);
 
     const ctx = canvas.getContext("2d");
     if (!ctx) return console.error("Could not get canvas context");
@@ -29,7 +19,7 @@
     ctx.beginPath();
 
     // if empty, just draw a line
-    if (dataArray.every((datum) => datum === 0)) {
+    if (data.every((datum) => datum === 0)) {
       ctx.moveTo(0, canvas.height / 2);
       ctx.lineTo(canvas.width, canvas.height / 2);
       ctx.stroke();
@@ -42,7 +32,7 @@
 
       // loop over step-sized chunks of data, find min and max, and draw to screen for that "pixel column"
       for (let j = 0; j < step; j++) {
-        const datum = dataArray[i * step + j];
+        const datum = data[i * step + j];
 
         if (datum < min) min = datum;
         if (datum > max) max = datum;
@@ -52,11 +42,13 @@
       ctx.lineTo(i, ((1 + max) * canvas.height) / 2);
     }
     ctx.stroke();
-  };
-
-  let canvas: HTMLCanvasElement;
+  });
 </script>
 
-<div class="w-full h-24 rounded-md bg-indigo-200">
-  <canvas class="w-full h-full" bind:this={canvas} />
+<div class="w-full h-full flex justify-center items-center">
+  {#await metadata.audioData}
+    <LoadIcon />
+  {:then}
+    <canvas class="w-full h-full" bind:this={canvas} />
+  {/await}
 </div>
