+++ src/lib/components/util/ResizeStalk.svelte
