+++ src/routes/+layout.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-  import MediaPool from "$lib/components/mediaPool/MediaPool.svelte";
-  import Player from "$lib/components/Player.svelte";
+  import MediaPool from "$lib/components/MediaPool.svelte";
+  import Player from "$lib/components/player/Player.svelte";
   import ResizeStalk from "$lib/components/ResizeStalk.svelte";
-  import Timeline from "$lib/components/timeline/Timeline.svelte";
+  import Timeline from "$lib/components/Timeline.svelte";
   import { studio } from "$lib/stores";
   import "../app.css";
